PR Changes
Refactor MRECs to MRecs - Java
