+++ AppLovin MAX Demo App - Java/app/src/main/AndroidManifest.xml
@@ -105,13 +105,13 @@
             android:name="com.applovin.enterprise.apps.demoapp.ads.applovin.leaders.LeaderLayoutEditorActivity"
             android:label="@string/title_activity_leader_layout_editor" />
         <activity
-            android:name="com.applovin.enterprise.apps.demoapp.ads.applovin.mrecs.MrecDemoMenuActivity"
+            android:name="com.applovin.enterprise.apps.demoapp.ads.applovin.mrecs.MRecDemoMenuActivity"
             android:label="@string/title_activity_mrec_list" />
         <activity
-            android:name="com.applovin.enterprise.apps.demoapp.ads.applovin.mrecs.MrecProgrammaticActivity"
+            android:name="com.applovin.enterprise.apps.demoapp.ads.applovin.mrecs.MRecProgrammaticActivity"
             android:label="@string/title_activity_mrec_programmatic" />
         <activity
-            android:name="com.applovin.enterprise.apps.demoapp.ads.applovin.mrecs.MrecLayoutEditorActivity"
+            android:name="com.applovin.enterprise.apps.demoapp.ads.applovin.mrecs.MRecLayoutEditorActivity"
             android:label="@string/title_activity_mrec_layout_editor" />
         <activity
             android:name="com.applovin.enterprise.apps.demoapp.ads.applovin.eventtracking.EventTrackingActivity"
