+++ AppLovin MAX Demo App - Java/app/src/main/java/com/applovin/enterprise/apps/demoapp/ads/applovin/mrecs/MRecDemoMenuActivity.java
@@ -5,15 +5,15 @@ import android.content.Intent;
 import com.applovin.enterprise.apps.demoapp.ads.applovin.DemoMenuActivity;
 import com.applovin.enterprise.apps.demoapp.data.main.DemoMenuItem;
 
-public class MrecDemoMenuActivity
+public class MRecDemoMenuActivity
     extends DemoMenuActivity
 {
     @Override
     protected DemoMenuItem[] getListViewContents()
     {
         return new DemoMenuItem[] {
-                new DemoMenuItem( "Programmatic", new Intent( this, MrecProgrammaticActivity.class ) ),
-                new DemoMenuItem( "Layout Editor", new Intent( this, MrecLayoutEditorActivity.class ) ),
+                new DemoMenuItem( "Programmatic", new Intent( this, MRecProgrammaticActivity.class ) ),
+                new DemoMenuItem( "Layout Editor", new Intent( this, MRecLayoutEditorActivity.class ) ),
         };
     }
 }
